pyproject.toml
[project]
name = "baseline"
version = "0.1.0"
description = "A baseline backend project."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.30.0",
    "alembic>=1.18.4",
    "email-validator>=2.2.0",
    "fastapi>=0.132.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "psycopg[binary]>=3.3.3",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.41.0"
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pylint>=4.0.5",
    "ruff>=0.15.2",
    "types-passlib>=1.7.7",
]

[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[tool.hatch.build.targets.wheel]
packages = ["src/baseline"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".eggs",
    ".git",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4
fix = true
show-fixes = true

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["C408"]
fixable = ["ALL"]
unfixable = ["F401"]  # Don't auto-fix unused imports

# B008 (no function calls in defaults) is a false positive for FastAPI Depends()
[tool.ruff.lint.per-file-ignores]
"src/baseline/api/*.py" = ["B008"]

[tool.ruff.format]
docstring-code-line-length = 60
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
namespace_packages = true
